Improve tests for the instrument_manager module.
